Adding 3 icons to Navigation Bar
