+++ recipE-Z/App.js
@@ -10,10 +10,10 @@ import { AppLoading } from 'expo';
 //imports of screens
 
 import StartScreen from './src/screens/StartScreen';
-import HomeScreen from './src/screens/HomeScreen';
 import ProfileScreen from './src/screens/ProfileScreen';
 import SignupScreen from './src/screens/SignupScreen';
 import LoginScreen from './src/screens/LoginScreen';
+import ResultsScreen from './src/screens/ResultsScreen';
 import UploadRecipeScreen from './src/screens/UploadRecipeScreen';
 import RecipeBookScreen from './src/screens/RecipeBookScreen';
 import SearchScreen from './src/screens/SearchScreen';
@@ -61,21 +61,21 @@ export default class App extends React.Component {
     });
     this.setState({ fontLoaded: true });
   }
-  
+
   render() {
   return (
     this.state.fontLoaded ? (
     <NavigationContainer>
       <Stack.Navigator initialRouteName="Start">
         <Stack.Screen name="Start" component={StartScreen} />
-        <Stack.Screen name="Home" component={HomeScreen} />
         <Stack.Screen name="Profile" component={ProfileScreen} />
         <Stack.Screen name="Sign-Up" component={SignupScreen} />
         <Stack.Screen name="Log-In" component={LoginScreen} />
         <Stack.Screen name="Upload" component={UploadRecipeScreen} />
-        <Stack.Screen name="Recipe Book" component={RecipeBookScreen} />
+        <Stack.Screen name="RecipeBook" component={RecipeBookScreen} />
         <Stack.Screen name="Search" component={SearchScreen} />
         <Stack.Screen name="Recipe" component={RecipeScreen} />
+        <Stack.Screen name="Results" component={ResultsScreen} />
       </Stack.Navigator>
     </NavigationContainer>
     ) : null
