+++ recipE-Z/src/screens/LoginScreen.js
@@ -65,7 +65,7 @@ class LoginScreen extends React.Component {
       </TouchableOpacity>
       <TouchableOpacity
         style={[defaultStyles.redButton, {}]}
-        onPress={() => navigation.navigate('Home')}
+        onPress={() => navigation.navigate('Search')}
         underlayColor='#ed4848'>
         <Text style={defaultStyles.redButtonText}>LOGIN</Text>
       </TouchableOpacity>
